+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Added
 
  - Add provider configuration for non default workflow types
+ - Show success column in state history overview
  
 ### Changed